+++ core/src/annostorage/inmemory.rs
@@ -720,7 +720,6 @@ where
                 if let Some(anno_key) = self.anno_keys.get_symbol(&anno_key) {
                     if let Some(histo) = self.histogram_bounds.get(&anno_key) {
                         // find the range in which the value is contained
-
                         // we need to make sure the histogram is not empty -> should have at least two bounds
                         if histo.len() >= 2 {
                             sum_histogram_buckets += histo.len() - 1;
@@ -752,6 +751,10 @@ where
     fn guess_max_count_regex(&self, ns: Option<&str>, name: &str, pattern: &str) -> Result<usize> {
         let full_match_pattern = util::regex_full_match(pattern);
 
+        // Get the total number of annotations with the namespace/name. We
+        // can't get larger than this number
+        let total = self.number_of_annotations_by_name(ns, name)?;
+
         // Try to parse the regular expression
         let parsed = regex_syntax::Parser::new().parse(&full_match_pattern);
         if let Ok(parsed) = parsed {
@@ -770,11 +773,13 @@ where
                         guessed_count += self.guess_max_count(ns, name, lower_val, &upper_val)?;
                     }
                 }
+            } else {
+                // For regular expressions without a prefix the worst case would be `.*[X].*` where `[X]` are the most common characters.
+                // Assume that a generic percentage (here 5%) of all nodes match the regex.
+                // TODO: find better ways of estimating this constant
+                guessed_count = (0.05 * (total as f64)) as usize;
             }
 
-            // Get the total number of annotations with the namespace/name. We
-            // can't get larger than this number
-            let total = self.number_of_annotations_by_name(ns, name)?;
             Ok(guessed_count.min(total))
         } else {
             Ok(0)