+++ core/src/annostorage/ondisk.rs
@@ -895,6 +895,10 @@ where
     fn guess_max_count_regex(&self, ns: Option<&str>, name: &str, pattern: &str) -> Result<usize> {
         let full_match_pattern = util::regex_full_match(pattern);
 
+        // Get the total number of annotations with the namespace/name. We
+        // can't get larger than this number
+        let total = self.number_of_annotations_by_name(ns, name)?;
+
         // Try to parse the regular expression
         let parsed = Parser::new().parse(&full_match_pattern);
         if let Ok(parsed) = parsed {
@@ -913,11 +917,13 @@ where
                         guessed_count += self.guess_max_count(ns, name, lower_val, &upper_val)?;
                     }
                 }
+            } else {
+                // For regular expressions without a prefix the worst case would be `.*[X].*` where `[X]` are the most common characters.
+                // Assume that a generic percentage (here 5%) of all nodes match the regex.
+                // TODO: find better ways of estimating this constant
+                guessed_count = (0.05 * (total as f64)) as usize;
             }
 
-            // Get the total number of annotations with the namespace/name. We
-            // can't get larger than this number
-            let total = self.number_of_annotations_by_name(ns, name)?;
             Ok(guessed_count.min(total))
         } else {
             Ok(0)