+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New optional `file` option for the `[logging]` section in the webservice
 configuration. Can be used to additionally output all log messages to the given
 file.
+- Add number of root nodes to graph storage statistics. This changes the way
+most of the graph storages store their statistics. You can use old imported data
+files, but to make use of the new information you queries, you have to
+**reimport** your corpora.
 - `Graph:ensure_loaded_parallel` returns the actually loaded components that did
 exist.
 
@@ -23,6 +27,8 @@ exist.
 
 - Less frequent corpus cache status updates in log. Before, every corpus access
 could trigger an entry into the log which is not desired under heavy load.
+- Improve query execution planning by assuming all annotations can be matched in
+regular expressions without a prefix.
 
 ## [3.7.1] - 2025-04-14